fix: resolve all mypy type errors across 35 files
- Add [tool.mypy] config to pyproject.toml (ignore_missing_imports, exclude migrations/cli)
- Convert admin short_description attrs to @admin.display() decorators
- Fix union-typed dict access with separate variables or dict[str, Any]
- Fix list method shadowing builtin with builtins.list annotations
- Add missing has_permission to IsOwner/IsOwnerOrAdmin permission classes
- Add type: ignore[attr-defined] for RefreshToken.access_token dynamic attr
- Fix type narrowing with or "" defaults and None guards
- Add missing CONTROLLER_TEMPLATE/MODEL_TEMPLATE/SCHEMA_TEMPLATE constants
- Fix int/float type mismatches in db_dump and metrics
- Correct type: ignore error codes in settings files
- Fix callable -> Callable, date -> datetime return types in test utils
- Convert misplaced module docstring to comments in api/urls.py

Note: pre-commit mypy with django-stubs still skipped — 140+ model field
annotation errors from django-stubs are a separate scope.

+++ core/management/commands/db_dump.py
@@ -338,7 +338,7 @@ class Command(BaseCommand):
             self.style.SUCCESS(f"✅ Cleaned up {len(to_delete)} old dumps")
         )
 
-    def format_size(self, size: int) -> str:
+    def format_size(self, size: float) -> str:
         """Format file size in human-readable format."""
         for unit in ["B", "KB", "MB", "GB"]:
             if size < 1024:
